Added changes as suggested by Matthias Kurz for Solaris 2.7

+++ openslp/slpd/slpd_main.c
@@ -251,7 +251,7 @@ int Daemonize(const char* pidfile)
     fd = fopen(pidfile,"w");
     if(fd)
     {
-        sprintf(pidstr,"%i",getpid());
+        sprintf(pidstr,"%i",(int)getpid());
         fwrite(pidstr,strlen(pidstr),1,fd);
         fclose(fd);
     }
